refactor: 订单/成交/持仓数量统一为int类型，清理浮点精度防御代码
- Order.quantity/filled_quantity: float -> int
- Trade.quantity: float -> int
- PriceLevel.total_quantity: double -> int
- Position.quantity: double -> int

清理 MatchingEngine 中为避免浮点精度问题的防御性代码:
- 删除 MIN_QTY = 1e-9 阈值
- 删除 MAX_ITERATIONS/MAX_NO_PROGRESS 迭代限制
- 删除残单检测逻辑
- 简化完全成交判断

清理 Position 中的 eps = 1e-9 精度判断

+++ src/bio/agents/base.py
@@ -351,7 +351,7 @@ class Agent:
 
         return quantity
 
-    def _place_limit_order(self, side: OrderSide, price: float, quantity: float, event_bus: EventBus) -> None:
+    def _place_limit_order(self, side: OrderSide, price: float, quantity: int, event_bus: EventBus) -> None:
         """创建并发布限价单
 
         Args:
@@ -372,7 +372,7 @@ class Agent:
         event = Event(EventType.ORDER_PLACED, 0.0, {"order": order})
         event_bus.publish(event)
 
-    def _place_market_order(self, side: OrderSide, quantity: float, event_bus: EventBus) -> None:
+    def _place_market_order(self, side: OrderSide, quantity: int, event_bus: EventBus) -> None:
         """创建并发布市价单
 
         Args:
@@ -525,7 +525,7 @@ class Agent:
         self,
         side: OrderSide,
         price: float,
-        quantity: float,
+        quantity: int,
         matching_engine: "MatchingEngine",
     ) -> list[Trade]:
         """直接下限价单（训练模式）
@@ -565,7 +565,7 @@ class Agent:
     def _place_market_order_direct(
         self,
         side: OrderSide,
-        quantity: float,
+        quantity: int,
         matching_engine: "MatchingEngine",
     ) -> list[Trade]:
         """直接下市价单（训练模式）
